Catch and read off out of order packets
In some cases, the gateway will send us packets we didn't request,
such as the packet for when the weather forecast changes. Until
we're ready to deal with those, we instead read them off the socket
buffer and throw them away. Because the next packet buffer in line
is what we actually wanted.